+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Security
+- All keys are now checked on parsing from the underlying library
+
+### Deprecated
+- `(key *Key) Check() (bool, error)` is now deprecated, all keys are now checked upon import from x/crypto
+
 ## [2.2.1] 2021-07-27
 ### Changed
 - Changed the returned `SignatureVerificationError.Status` when trying to verify a message with no embedded signature. It used to return `constants.SIGNATURE_NO_VERIFIER` and now returns `constants.SIGNATURE_NOT_SIGNED`.