+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Security
 - All keys are now checked on parsing from the underlying library
 
+### Fixed
+- Dummy keys now show the correct locked/unlocked status
+
 ### Deprecated
 - `(key *Key) Check() (bool, error)` is now deprecated, all keys are now checked upon import from x/crypto